third_party/blink/web_tests/wpt_internal/css/css-overflow/parsing/scroll-marker-contain-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Overflow: parsing scroll-marker-contain property invalid values</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/10916">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<div id="target"></div>
<script>
  test_invalid_value('scroll-marker-contain', '10');
  test_invalid_value('scroll-marker-contain', 'true');
  test_invalid_value('scroll-marker-contain', 'default');
  test_invalid_value('scroll-marker-contain', 'all');
  test_invalid_value('scroll-marker-contain', 'auto, auto');
</script>
